Fix for accessing invalid memory in FMTTask_SerializeActors::DumpData()
In the previous implementation FMTTask_SerializeActors can be passed a pointer to a LevelRecord in the SubLevels array that is subsequently made invalid by calling SlotData->SubLevels.AddUnique(). This bug will occur if there are enough sublevels to require the SubLevels array to resize.

In the new implementation an id is used to store a reference to the level (either the array index of the level or -1 to indicate the persistent level) instead of a pointer.